# Optimized simulator trace contexts.